# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  detail?: string | null;
}

export interface UploadResult {
  session_id: string;
  filename: string;
  rows: number;
  columns: number;
  message: string;
  column_names?: string[];
}

export interface ColumnInfo {
  name: string;
  dtype: string;
  null_count: number;
  null_pct: number;
  sample_values?: unknown[];
}

export type DatasetSummaryNumericSummary = {[key: string]: unknown};

export interface DatasetSummary {
  session_id: string;
  filename: string;
  rows: number;
  columns: number;
  columns_info: ColumnInfo[];
  numeric_summary?: DatasetSummaryNumericSummary;
  /** @nullable */
  memory_usage?: string | null;
}

export interface DatasetPreview {
  columns: string[];
  rows: unknown[][];
  total_rows: number;
}

export interface QueryInput {
  /** @minLength 1 */
  question: string;
  /** @nullable */
  session_id?: string | null;
}

export type ChartDataType = typeof ChartDataType[keyof typeof ChartDataType];


export const ChartDataType = {
  bar: 'bar',
  line: 'line',
  pie: 'pie',
  scatter: 'scatter',
  histogram: 'histogram',
  none: 'none',
} as const;

export interface ChartData {
  type: ChartDataType;
  /**
     * Plotly figure serialized as JSON string
     * @nullable
     */
  plotly_json?: string | null;
}

export interface QueryResult {
  query_id: string;
  question: string;
  answer: string;
  chart?: ChartData;
  result_table?: DatasetPreview;
  insights?: string[];
  /** @nullable */
  generated_code?: string | null;
  execution_time_ms?: number;
}

export interface QueryHistoryItem {
  query_id: string;
  question: string;
  answer: string;
  /** @nullable */
  chart_type?: string | null;
  created_at: string;
  has_chart?: boolean;
  has_table?: boolean;
}

export interface ExportInput {
  query_id: string;
}

export interface SampleDataset {
  id: string;
  name: string;
  description: string;
  rows: number;
  columns: number;
}

export interface SampleDatasetInput {
  dataset_id: string;
}

export type GetDatasetSummaryParams = {
/**
 * Session ID for the dataset
 */
session_id?: string;
};

export type GetDatasetPreviewParams = {
session_id?: string;
limit?: number;
};

export type GetQueryHistoryParams = {
session_id?: string;
};