# lib/api-zod/src/generated/types/chartData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartDataType } from './chartDataType';

export interface ChartData {
  type: ChartDataType;
  /**
     * Plotly figure serialized as JSON string
     * @nullable
     */
  plotly_json?: string | null;
}