# lib/api-zod/src/generated/types/datasetPreview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface DatasetPreview {
  columns: string[];
  rows: unknown[][];
  total_rows: number;
}