# lib/api-zod/src/generated/types/datasetSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnInfo } from './columnInfo';
import type { DatasetSummaryNumericSummary } from './datasetSummaryNumericSummary';

export interface DatasetSummary {
  session_id: string;
  filename: string;
  rows: number;
  columns: number;
  columns_info: ColumnInfo[];
  numeric_summary?: DatasetSummaryNumericSummary;
  /** @nullable */
  memory_usage?: string | null;
}