# lib/api-zod/src/generated/types/queryHistoryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface QueryHistoryItem {
  query_id: string;
  question: string;
  answer: string;
  /** @nullable */
  chart_type?: string | null;
  created_at: string;
  has_chart?: boolean;
  has_table?: boolean;
}