# lib/api-zod/src/generated/types/queryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface QueryInput {
  /** @minLength 1 */
  question: string;
  /** @nullable */
  session_id?: string | null;
}