# lib/api-zod/src/generated/types/queryResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */
import type { ChartData } from './chartData';
import type { DatasetPreview } from './datasetPreview';

export interface QueryResult {
  query_id: string;
  question: string;
  answer: string;
  chart?: ChartData;
  result_table?: DatasetPreview;
  insights?: string[];
  /** @nullable */
  generated_code?: string | null;
  execution_time_ms?: number;
}