# lib/api-zod/src/generated/types/sampleDataset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface SampleDataset {
  id: string;
  name: string;
  description: string;
  rows: number;
  columns: number;
}