# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CSV Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  session_id: string;
  filename: string;
  rows: number;
  columns: number;
  message: string;
  column_names?: string[];
}